Revert "Ran sed script to convert real litterals to floats."
This reverts commit d7e31e9566f43c54dcb1534c9ee1fef273046f54.

+++ release1.0/src/Common.h
@@ -238,7 +238,7 @@ string static GetCurrentDir(void) {
 class Common
 {
 public:
-	float static estar(float t);
+	double static estar(float t);
 	static float Qs(float P, float tc);
 	float static Qswater(float P, float tc);
 	float static DepthofSnow(float SWE);
